Delete Blog Post URL Scanner.ipynb
a python scanner that takes in URL of cluster of posts (i.e. tags), and scan to see whether there is a specific URL as hyperlink in it. This can be easily edited to be scanning for a particular word or phrase. Particularly useful for content management in content heavy drupal site.